Update CORS settings and refactor job creation function
This commit updates CORS settings to define accepted methods explicitly and allow any origin.

Changes in job-related scripts were also made. The method of creating jobs was refactored: the previous version was depending on the req, res parameters, but now it solely depends on the jobData, userId parameters which leads to more flexible usage.

A user_id field was also added to the AddJobs component to associate the current logged-in user with the job that's being created. The user_id is passed from the frontend to the create job service through the body of the POST request.

Furthermore, adjustment in the job routes was made for a better managing of jobs for a specific user. With the revised route, it's much clearer that a job is being created for a particular user.

The changes would improve code maintainability and the user's experience on managing job postings.