Modify application status message on job cards
Changed the message displayed on job cards from 'Click to Apply' to 'If Applied Click Here'. This change was made to clarify the status of applications and reduce possible confusion for users. The new message effectively communicates that a click is required only if the user had applied for the job.