Redirect to login if job list is empty
Added a condition to redirect users to the login page if there are no jobs in the jobs array. This is to ensure that only logged-in users can see the job list. In the job tracker components, we also fixed a potential type error by explicitly handling empty job lists in Jumbo.js and Cards.js.