Refactor job tracker codes and add a new file for job application check
Work has been done to refactor the job tracker codes. The handleCheck function in Cards.js was removed as it was replaced with the IsApplied.js file. The latter contains the function for checking the status of job applications. This was done to improve the organization of the code and its maintain