Refactor code to remove duplication and enhance readability
The change mainly refactors few files within the jobtracker component. Common URL and headers are extracted out to remove duplication and improve code maintainability. fetchJobs.js was deleted and its logic was moved into Cards.js as it was only used in that file. This improves overall clarity and organization of the code.