Add new components and refactor Jumbo.js for better organization
This commit involves adding new components for the job tracker feature. Components for the DarkModeButton, the JobCarousel, the LogInRedirect, the AddJobButton, and the JumbotronBackground are initialized. existing Jumbo.js file is refactored to import and use these new components for increased modularity and organization. In Jumbo.js, the function fetchJobs is moved inside the useEffect scope for better encapsulation, and replaced multiple instances where ShowModal's state was being toggled separately with a single handleModalToggle function. These changes aim to better manage state changes and make the codebase more maintainable.