Add ScrollToTop feature and react-icons dependency
Added a new ScrollToTop feature to improve user navigation in the JobTracker page. It displays an up-arrow button to take the user to the top of the page when they scroll past a certain point. For this feature, a new dependency 'react-icons' was added to the project.

Also made minor enhancements like changing the color of the 'Show More' button and adding react-icons to the package and package-lock.json files.