Add "Scroll to Top" button and modify "Cards" component
Integrated a "Scroll to top" feature on the Jobtracker app page for improved user navigation, especially during long scrolls. This necessitated creating new css file 'scrollToTop.css' and 'ScrollToTop.js' files. The scroll to top feature detects whether the user has scrolled down and accordingly shows the button on a fixed position.

In addition, the 'Cards' component was modified to include a "Show More" button to limit the number of jobs shown at once and improve the loading times during initial render. The backend API URL was also changed to