Reduce jumbotron height and refactor news ticker display
Adjusted the height of the `JumbotronBackground` for better visual balance within the layout of JobTracker. In addition, the margin settings in `Jumbo.js` have been revised for better alignment. The news ticker in `NewsTicker.css` and `NewsTicker.js` has been modified to slow down the display speed for better readability, and pause the ticker animation on hover for user-friendly interaction. Also added functionality to fetch news headlines from an external API source, displaying headline title with a corresponding link.