Add Stocks Ticker and adjust styling
A new Stocks component is created to display real-time changes in stocks, forex, and bitcoin prices. The component fetches the data from an API and displays it in a ticker layout. With this feature, users can monitor current financial market trends without leaving the application.

Other changes include the adjustment of the position property of several elements from "fixed" to "relative" for better layout consistency. Database connection configurations in db.config.js are also updated with "localhost" fallbacks for the ease of local testing.

The JumbotronBackground.js and Pages.js now import and utilize the Stocks component. A new file, Stocks.css, is also added to handle the styling of the financial data ticker.